+++ src/Illuminate/Cache/FileStore.php
@@ -102,7 +102,7 @@ class FileStore implements Store, LockProvider
 
         try {
             $file->getExclusiveLock();
-        } catch (LockTimeoutException $e) {
+        } catch (LockTimeoutException) {
             $file->close();
 
             return false;
@@ -279,7 +279,7 @@ class FileStore implements Store, LockProvider
             $expire = substr(
                 $contents = $this->files->get($path, true), 0, 10
             );
-        } catch (Exception $e) {
+        } catch (Exception) {
             return $this->emptyPayload();
         }
 
@@ -294,7 +294,7 @@ class FileStore implements Store, LockProvider
 
         try {
             $data = unserialize(substr($contents, 10));
-        } catch (Exception $e) {
+        } catch (Exception) {
             $this->forget($key);
 
             return $this->emptyPayload();