+++ src/Illuminate/Foundation/Bootstrap/HandleExceptions.php
@@ -90,7 +90,7 @@ class HandleExceptions
 
         try {
             $logger = static::$app->make(LogManager::class);
-        } catch (Exception $e) {
+        } catch (Exception) {
             return;
         }
 
@@ -180,7 +180,7 @@ class HandleExceptions
 
         try {
             $this->getExceptionHandler()->report($e);
-        } catch (Exception $e) {
+        } catch (Exception) {
             //
         }