+++ src/Illuminate/Foundation/Support/Providers/AuthServiceProvider.php
@@ -19,6 +19,18 @@ class AuthServiceProvider extends ServiceProvider
      *
      * @return void
      */
+    public function register()
+    {
+        $this->booting(function () {
+            $this->registerPolicies();
+        });
+    }
+
+    /**
+     * Register the application's policies.
+     *
+     * @return void
+     */
     public function registerPolicies()
     {
         foreach ($this->policies() as $model => $policy) {