+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for 10.x
 
-## [Unreleased](https://github.com/laravel/framework/compare/v10.0.1..10.x)
+## [Unreleased](https://github.com/laravel/framework/compare/v10.0.1...10.x)
 
 
 ## [v10.0.1 (2023-02-16)](https://github.com/laravel/framework/compare/v10.0.0...v10.0.1)
@@ -9,12 +9,11 @@
 - Standard Input can be applied to PendingProcess ([#46119](https://github.com/laravel/framework/pull/46119))
 
 ### Fixed
-- Fix Expression string casting ([#4613](https://github.com/laravel/framework/pull/4613))
+- Fix Expression string casting ([#46137](https://github.com/laravel/framework/pull/46137))
 
 ### Changed
 - Add AddQueuedCookiesToResponse to middlewarePriority so it is handled in the right place ([#46130](https://github.com/laravel/framework/pull/46130))
 - Show queue connection in MonitorCommand ([#46122](https://github.com/laravel/framework/pull/46122))
-
 
 ## [v10.0.0 (2023-02-14)](https://github.com/laravel/framework/compare/v10.0.0...10.x)